+++ src/pages/Kusujeong/Main/Comment.js
@@ -1,14 +1,16 @@
 import React from 'react';
 
-export default function Comment({ userName, comment, key }) {
+export default function Comment({ comment, deleteComment }) {
   return (
-    <li className="description-comment" key={key}>
+    <li className="description-comment">
       <div className="article-info">
-        <div className="ariticle-id">{userName}</div>
+        <div className="ariticle-id">rntnwjd</div>
         <div className="article-description">{comment}</div>
       </div>
       <div className="comment-function">
-        <button className="comment-delete">삭제</button>
+        <button className="comment-delete" onClick={deleteComment}>
+          삭제
+        </button>
         <i className="fa-regular fa-heart" />
       </div>
     </li>
