+++ src/pages/Kusujeong/Main/MainKsj.js
@@ -1,34 +1,14 @@
 import React, { useState, useEffect } from 'react';
 import Nav from '../../../components/Nav/Nav';
-import Comment from './Comment';
+import Feed from './Feed';
 import { FOOTER_INFO_LIST } from './Footer';
 import './MainKsj.scss';
 
 export default function MainKsj() {
   const imgUrl = '/images/Kusujeong/IMG_8905.JPG';
 
-  // Input창에 입력한 텍스트 저장할 state
-  const [comment, setComment] = useState('');
-  // 댓글을 저장할 배열
-  const [commentList, setCommentList] = useState([]);
   // mock data에서 불러온 피드 정보
   const [feedList, setFeedList] = useState([]);
-  // 엔터나 게시버튼을 누르면 commentList에 입력한 댓글을 저장하고 그걸 html에 뿌려준다
-
-  // 20221108 피드 별 댓글 기능 구현중 - 구수정
-  const regiComment = e => {
-    if (e.key === 'Enter' || e.target.nodeName === 'BUTTON') {
-      if (comment.length > 0) {
-        const commentObj = {
-          id: 1,
-          userName: 'rntnwjd',
-          commentContent: comment,
-        };
-        setFeedList([...feedList.comments, commentObj]);
-        setComment('');
-      }
-    }
-  };
 
   useEffect(() => {
     fetch('/data/feedData.json')
@@ -97,89 +77,7 @@ export default function MainKsj() {
           </div>
           {/* <!-- 피드 --> */}
           {feedList.map(feed => (
-            <div className="feed" key={feed.id}>
-              <div className="feed-header">
-                <div className="feed-logo">
-                  <img
-                    className="feed-profile"
-                    src={feed.profileImg}
-                    alt="피드 프로필이미지"
-                  />
-                  {feed.userName}
-                </div>
-                <i className="fa-solid fa-ellipsis feed-menu" />
-              </div>
-              <div className="feed-article">
-                <img className="feed-img" src={feed.feedImg} alt="피드 사진" />
-              </div>
-              <div className="feed-list">
-                <div className="feed-left_list">
-                  <i className="fa-regular fa-heart" />
-                  <i className="fa-regular fa-comment" />
-                  <i className="fa-regular fa-paper-plane" />
-                </div>
-                <div className="feed-middle_list">
-                  <span className="middle-dot" />
-                  <span className="middle-dot" />
-                  <span className="middle-dot" />
-                  <span className="middle-dot" />
-                  <span className="middle-dot" />
-                  <span className="middle-dot" />
-                </div>
-                <div className="feed-right_list">
-                  <i className="fa-regular fa-bookmark" />
-                </div>
-              </div>
-              <div className="feed-description">
-                <div className="description-like">좋아요 3,460개</div>
-                <div className="description-article">
-                  <div className="article-info">
-                    <div className="ariticle-id">{feed.userName}</div>
-                    <div className="article-description">
-                      {feed.feedContent}
-                    </div>
-                  </div>
-                  <div className="article-more">...더보기</div>
-                </div>
-                <div className="description-commentAll">
-                  댓글 {commentList.length}개 모두 보기
-                </div>
-                <ul id="commentList">
-                  {commentList.map((comment, index) => (
-                    <Comment
-                      userName={feed.comments[index].userName}
-                      comment={comment}
-                      key={feed.comments[index].id}
-                    />
-                  ))}
-                </ul>
-              </div>
-              <div className="description-date">2일 전</div>
-              <div className="feed-footer">
-                <div className="footer-logo">
-                  <div className="footer-smile">
-                    <i className="fa-regular fa-face-smile" />
-                  </div>
-                  <div className="footer-comment">
-                    <input
-                      className="comment-input"
-                      type="text"
-                      placeholder="댓글달기"
-                      onChange={e => {
-                        setComment(e.target.value);
-                      }}
-                      onKeyUp={regiComment}
-                      value={comment}
-                    />
-                  </div>
-                </div>
-                <div className="footer-enter">
-                  <button className="comment-btn btn" onClick={regiComment}>
-                    게시
-                  </button>
-                </div>
-              </div>
-            </div>
+            <Feed feedInfo={feed} key={feed.id} />
           ))}
         </article>
         <article className="main-right">
